+++ NEWS.md
@@ -1,6 +1,8 @@
 # Version 0.1.4 [developtment]
 ### Minor changes
 * Include complexity meta-features
+* Include concept characterization meta-features
+* Include itemset meta-features
 * Include the option "default" to distinguish from "all" meta-features groups
 
 # Version 0.1.3 [current]